Add files via upload
Here's how to create a simple QR Code Generator using HTML, CSS, and JavaScript. This QR Code generator will allow users to input text and generate a corresponding QR code.
To generate QR codes, we will use the QRCode.js library, which is easy to integrate and use in the web app.
How it works:
HTML provides a basic form with an input field for entering text or URLs and a button to generate the QR code.
CSS styles the page, ensuring the form is centered and visually appealing.
JavaScript uses the QRCode.js library to generate the QR code. When the user inputs text and clicks the button, the generateQRCode() function is triggered, creating the QR code and displaying it in the specified div.
Steps to use:
Save the code in three separate files: index.html, styles.css, and script.js.
Open the index.html file in a browser.
Enter some text or a URL in the input field and click the "Generate QR Code" button to generate the QR code.
Libraries Used:
The QRCode.js library is included from a CDN link to generate the QR code. You can replace this with a local version if you prefer.
This is a simple and functional QR Code generator web app. You can further customize it or add features such as the ability to download the generated QR code.